+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+
+
 ## 1.2.3
 
 * Fix `b3` and `tracecontext` using incorrect Trace and Span ids during request tracking.